Quick refactor fixing all inconsistent references after the namespace name change. Added user attributes and onlineUsers attribute to server.
