+++ ClassLibrary/ClassLibrary.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Net.Sockets;
 
-namespace ClassLibrary
+namespace Protocol
 {
     public class ClassLibrary
     {
