+++ Domain/User.cs
@@ -4,9 +4,16 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ClassLibrary
+namespace Domain
 {
-    class User
+    public class User
     {
+        private string username;
+        private string password;
+        private List<User> friends;
+
+        public string Username { get => username; set => username = value; }
+        public string Password { get => password; set => password = value; }
+        internal List<User> Friends { get => friends; set => friends = value; }
     }
 }
