+++ Sockets/Server.csproj
@@ -54,6 +54,10 @@
       <Project>{EB71029B-3998-4881-9E31-83B69BFABD91}</Project>
       <Name>Protocol</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Domain\Domain.csproj">
+      <Project>{6D5B56FD-0CCC-43CE-AFA6-CC2314D6FD70}</Project>
+      <Name>Domain</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
