+++ Sockets/bin/Debug/Domain.pdb
